test: refactor test
